requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.0
vine==5.1.0
wcwidth==0.2.13
